requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
grpcio>=1.54.0
grpcio-tools>=1.54.0
protobuf>=4.23.0
psutil>=5.9.0
requests>=2.31.0
pydantic>=2.0.0
websockets>=11.0